agent_catalog_libs/pyproject.toml
[tool.poetry]
name = "agent_catalog_libs"
version = "0.1.0"
description = "Core library associated with agentic workflows in Couchbase."
repository = "https://github.com/couchbaselabs/rosetta-core"
authors = [
    "Glenn Galvizo <glenn.galvizo@couchbase.com>",
    "Steve Yen <steve.yen@couchbase.com>",
    "Tanvi Johari <tanvi.johari@couchbase.com>"
]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "cmd" },
    { include = "util" }
]

include = [
    # These are the Jinja templates we use for code-gen.
    "core/tool/generate/templates/*",

    # These are the Analytics DDLs for log analysis.
    "core/analytics/ddls/*"
]

[tool.poetry.urls]
Homepage = "https://couchbase.com"
Documentation = "https://docs.couchbase.com"
Repository = "https://github.com/couchbaselabs/rosetta-core"

[tool.poetry.scripts]
agentc = "cmd.main:main"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
PyYAML = "^6.0.1"
semantic_version = "^2.10.0"
gitignore-parser = "^0.1.11"
GitPython = "^3.1.43"
sentence-transformers = "^3.1.1"
pydantic = "^2.8.2"
datamodel-code-generator = "^0.26.0"
couchbase = "^4.3.0"
jinja2 = "^3.1.4"
openapi3-parser = "^1.1.17"
py-openapi-schema-to-json-schema = "^0.0.3"
jsbeautifier = "^1.15.1"

[tool.poetry.group.cli.dependencies]
click = "^8.1.7"
tqdm = "^4.66.5"

[tool.poetry.group.refiner.dependencies]
scikit-learn = ">1.0.2"
numpy = ">1.13.3"
scipy = ">0.14.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
pytest = "^8.3.2"

# Optional groups are specified below.
[tool.poetry.group.analyze]
optional = true

[tool.poetry.group.analyze.dependencies]
# Note: the langchain dependency specification is to speed up poetry.
langchain = "^0.2.16"
ragas = "^0.1.20"

[tool.poetry.group.web]
optional = true

[tool.poetry.group.web.dependencies]
Flask = "^3.0.3"

# TODO (GLENN): Can we factor out the controlflow dependency?
[tool.poetry.group.exec_cmd.dependencies]
controlflow = "^0.9.4"

[tool.pytest.ini_options]
markers = [
    "smoke: marks tests for running on each git-push"
]

[tool.ruff]
line-length = 120
lint.select = ["F", "B", "I", "SIM"]
show-fixes = true

[tool.ruff.lint.isort]
no-sections = true
force-single-line = true
lines-between-types = 1
force-wrap-aliases = true
split-on-trailing-comma = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
